pyproject.toml
[tool.poetry]
name = "google-calendar"
version = "0.1.0"
description = ""
authors = ["toma tanaka <ulti4929@gmail.com>"]
readme = "README.md"
packages = [{include = "google_calendar"}]

[tool.poetry.dependencies]
python = "^3.9.0, <3.13.0"
google-api-python-client = "1.7.2"
google-auth = "1.8.0"
oauthlib = "^3.2.2"
google-auth-oauthlib = "0.4.1"
google-auth-httplib2 = "0.0.3"
ipykernel = "^6.25.2"
jupyter = "^1.0.0"
notebook = "^7.0.4"
pandas = "^2.1.1"
numpy = "^1.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
